@@ -1,5 +1,10 @@
 module tlgstats
 
-go 1.14
+go 1.19
 
 require github.com/PuerkitoBio/goquery v1.5.0
+
+require (
+	github.com/andybalholm/cascadia v1.0.0 // indirect
+	golang.org/x/net v0.0.0-20181114220301-adae6a3d119a // indirect
+)
